Changed [withOpacity]'s parameter from [int] to [double].

+++ lib/src/color_model.dart
@@ -115,7 +115,7 @@ abstract class ColorModel {
   /// Returns this [ColorModel] with the provided [opacity] value.
   ///
   /// [opacity] is the equivalent of [alpha] normalized to a `0` to `1` value.
-  ColorModel withOpacity(int opacity);
+  ColorModel withOpacity(double opacity);
 
   /// Returns the distance between `this` color's
   /// hue and [color]'s hue in degrees.
